Extract more `dispute` tests
